+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `glsl-lang-cli`: Add input file argument
 - `glsl-lang-cli`: Add GLSL output format
 - `glsl-lang`: Implement PartialOrd and Hash for Identifier and TypeName
+- `lang-util`: Add NodeSpan::start() and NodeSpan::end()
 
 ### Changed
 - PartialEq for Node now compares the contents of the node by default